mock测试添加，测试通过

+++ JunitMockTest/src/main/java/com/code/junit/mock/boot/dict/dao/MockTableDAO.java
@@ -1,10 +1,6 @@
 package com.code.junit.mock.boot.dict.dao;
 
 import com.code.junit.mock.boot.dict.beans.MockTable;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
 
 /**
  * 映射mapper.xml文件操作数据库
@@ -17,10 +13,7 @@ import org.apache.ibatis.annotations.Update;
  */
 public interface MockTableDAO {
 
-//    @Insert("insert into mocktable (name,data) values( #{name},#{data})")
     public Long save(MockTable mockTable);
-//    @Select("select * from mocktable where id=#{id}")
-    public MockTable selectById(Integer id);
-//    @Update("update mocktable set name=#{name},data=#{data} where id=#{id}")
+    public MockTable selectById(Long id);
     public int updateById(MockTable mockTable);
 }
